pyproject.toml
[project]
name = "qrgen"
version = "0.1.0"
description = "A command-line tool to generate customizable QR codes"
authors = [
    {name = "Alex Eze",email = "ialexeze@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "qrcode (>=8.2,<9.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
qrgen = "qrgen.main:entry_point"
